Return a size_t from used_bytes

+++ include/walnut/big_int_base.h
@@ -35,7 +35,7 @@ class BigIntBase {
     return words_[0].ToInt();
   }
 
-  constexpr int used_bytes() const {
+  constexpr size_t used_bytes() const {
     return used_;
   }
